# swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Diagnoxix API",
        "title": "Diagnoxix",
        "contact": {},
        "version": "1.0"
    },
    "host": "diagnoxix.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/v1/diagnostic_centres": {
            "get": {
                "description": "Search for diagnostic centres based on location, available doctors, and test types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticCentre"
                ],
                "summary": "Search for diagnostic centres",
                "parameters": [
                    {
                        "maximum": 90,
                        "minimum": -90,
                        "type": "number",
                        "default": 25.06,
                        "description": "Latitude (-90 to 90)",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 180,
                        "minimum": -180,
                        "type": "number",
                        "default": 56.67,
                        "description": "Longitude (-180 to 180)",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "monday",
                            "tuesday",
                            "wednesday",
                            "thursday",
                            "friday",
                            "saturday",
                            "sunday"
                        ],
                        "type": "string",
                        "description": "Filter by day",
                        "name": "day_of_week",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "Male",
                            "Female"
                        ],
                        "type": "string",
                        "description": "Filter by doctor specialization",
                        "name": "doctor",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "BLOOD_TEST",
                            "URINE_TEST",
                            "X_RAY",
                            "MRI",
                            "CT_SCAN",
                            "ULTRASOUND",
                            "ECG",
                            "EEG",
                            "BIOPSY",
                            "SKIN_TEST",
                            "ALLERGY_TEST",
                            "GENETIC_TEST",
                            "IMMUNOLOGY_TEST",
                            "HORMONE_TEST",
                            "VIRAL_TEST",
                            "BACTERIAL_TEST",
                            "PARASITIC_TEST",
                            "FUNGAL_TEST",
                            "MOLECULAR_TEST",
                            "TOXICOLOGY_TEST",
                            "ECHO",
                            "COVID_19_TEST",
                            "OTHER",
                            "BLOOD_SUGAR_TEST",
                            "LIPID_PROFILE",
                            "HEMOGLOBIN_TEST",
                            "THYROID_TEST",
                            "LIVER_FUNCTION_TEST",
                            "KIDNEY_FUNCTION_TEST",
                            "URIC_ACID_TEST",
                            "VITAMIN_D_TEST",
                            "VITAMIN_B12_TEST",
                            "HEMOGRAM",
                            "COMPLETE_BLOOD_COUNT",
                            "BLOOD_GROUPING",
                            "HEPATITIS_B_TEST",
                            "HEPATITIS_C_TEST",
                            "HIV_TEST",
                            "MALARIA_TEST",
                            "DENGUE_TEST",
                            "TYPHOID_TEST",
                            "COVID_19_ANTIBODY_TEST",
                            "COVID_19_RAPID_ANTIGEN_TEST",
                            "COVID_19_RT_PCR_TEST",
                            "PREGNANCY_TEST"
                        ],
                        "type": "string",
                        "description": "Filter by available test type",
                        "name": "available_tests",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of diagnostic centres",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.DiagnosticCentreSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new diagnostic centre with location, contact details, and available services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticCentre"
                ],
                "summary": "Create a new diagnostic centre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Diagnostic centre details",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateDiagnosticDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Diagnostic centre created successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.DiagnosticCentreSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/diagnostic_centres/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticCentre"
                ],
                "summary": "Assign manager to a centre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Assign Manager to a diagnostic centre",
                        "name": "diagnostic_centre",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateDiagnosticManagerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Manager assigned successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.ManagerSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/diagnostic_centres/manager": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all diagnostic centres managed by the authenticated diagnostic centre manager",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticCentre"
                ],
                "summary": "List manager's diagnostic centres",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 50,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of diagnostic centres",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.DiagnosticCentreSwagger"
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new diagnostic centre manager account. Only accessible by diagnostic centre owners.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticCentre"
                ],
                "summary": "Create a diagnostic centre manager",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Manager details",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.DiagnosticCentreManagerRegisterDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Manager account created successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.ManagerSwagger"
                        }
                    },
                    "202": {
                        "description": "Manager invite sent successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "409": {
                        "description": "DUPLICATE_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.DUPLICATE_ERROR"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNPROCESSED_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.UpdateDiagnosticManagerDTO": {
            "type": "object",
            "required": [
                "diagnostic_centre_id",
                "manager_id"
            ],
            "properties": {
                "diagnostic_centre_id": {
                    "type": "string"
                },
                "manager_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.BAD_REQUEST": {
            "description": "BAD_REQUEST response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "BAD_REQUEST"
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.DiagnosticCentreSwagger": {
            "description": "Diagnostic centre response for Swagger",
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "123 Main St, Lagos"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-06-26T20:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "info@diagnoxix.com"
                },
                "id": {
                    "type": "string",
                    "example": "dc-001"
                },
                "latitude": {
                    "type": "number",
                    "example": 6.5244
                },
                "longitude": {
                    "type": "number",
                    "example": 3.3792
                },
                "name": {
                    "type": "string",
                    "example": "Diagnoxix Diagnostics"
                },
                "phone": {
                    "type": "string",
                    "example": "+2348000000000"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-06-26T20:30:00Z"
                }
            }
        },
...
    }
}